refactor: moves some files into a more appropriate place
