+++ src/app/common/EmptyBlock.vue
