+++ src/app/common/ErrorBlock.vue
