+++ src/app/common/LoadingBlock.vue
