+++ src/app/common/LoadingBox.vue
