+++ src/app/data-planes/views/DataPlaneDetailView.vue
@@ -25,9 +25,9 @@ import { DataPlane, DataPlaneOverview } from '@/types'
 import { useStore } from '@/store/store'
 import Kuma from '@/services/kuma'
 import DataPlaneDetails from '../components/DataPlaneDetails.vue'
-import EmptyBlock from '@/components/EmptyBlock.vue'
-import ErrorBlock from '@/components/ErrorBlock.vue'
-import LoadingBlock from '@/components/LoadingBlock.vue'
+import EmptyBlock from '@/app/common/EmptyBlock.vue'
+import ErrorBlock from '@/app/common/ErrorBlock.vue'
+import LoadingBlock from '@/app/common/LoadingBlock.vue'
 
 const route = useRoute()
 const store = useStore()
