+++ src/app/data-planes/views/DataPlaneListView.vue
@@ -125,7 +125,7 @@ import { KButton, KDropdownItem, KDropdownMenu } from '@kong/kongponents'
 import ContentWrapper from '@/app/common/ContentWrapper.vue'
 import DataOverview from '@/components/Skeletons/DataOverview.vue'
 import DataPlaneEntitySummary from '@/app/data-planes/components/DataPlaneEntitySummary.vue'
-import EmptyBlock from '@/components/EmptyBlock.vue'
+import EmptyBlock from '@/app/common/EmptyBlock.vue'
 import { columnsDropdownItems, defaultVisibleTableHeaderKeys, getDataPlaneTableHeaders } from '../constants'
 import { useStore } from '@/store/store'
 import { ClientStorage } from '@/utils/ClientStorage'
