+++ src/app/mesh-overview/views/MeshOverviewView.vue
@@ -104,7 +104,7 @@ import MeshCharts from '../components/MeshCharts.vue'
 import MeshResources from '@/app/common/MeshResources.vue'
 import LabelList from '@/components/Utils/LabelList.vue'
 import TabsWidget from '@/components/Utils/TabsWidget.vue'
-import YamlView from '@/components/Skeletons/YamlView.vue'
+import YamlView from '@/app/common/YamlView.vue'
 import Kuma from '@/services/kuma'
 import { Mesh, MeshInsight } from '@/types'
 import { humanReadableDate, stripTimes } from '@/helpers'
