+++ src/app/services/components/ServiceInsightDetails.vue
@@ -63,7 +63,7 @@ import { ServiceInsight } from '@/types'
 import { stripTimes } from '@/helpers'
 import { STATUS } from '@/consts'
 import EntityURLControl from '@/components/Utils/EntityURLControl.vue'
-import YamlView from '@/components/Skeletons/YamlView.vue'
+import YamlView from '@/app/common/YamlView.vue'
 
 const route = useRoute()
