+++ src/components/Skeletons/DataOverview.vue
@@ -264,9 +264,9 @@ import { datadogLogs } from '@datadog/browser-logs'
 
 import { datadogLogEvents } from '@/datadogEvents'
 import PaginationWidget from '@/components/PaginationWidget.vue'
-import EmptyBlock from '@/components/EmptyBlock.vue'
-import ErrorBlock from '@/components/ErrorBlock.vue'
-import LoadingBlock from '@/components/LoadingBlock.vue'
+import EmptyBlock from '@/app/common/EmptyBlock.vue'
+import ErrorBlock from '@/app/common/ErrorBlock.vue'
+import LoadingBlock from '@/app/common/LoadingBlock.vue'
 
 export default {
   name: 'DataOverview',
