+++ src/views/Entities/ZonesView.vue
@@ -126,7 +126,7 @@ import { mapGetters } from 'vuex'
 import SubscriptionDetails from './components/SubscriptionDetails.vue'
 import SubscriptionHeader from './components/SubscriptionHeader.vue'
 import MultizoneInfo from './components/MultizoneInfo.vue'
-import CodeBlock from '@/components/CodeBlock.vue'
+import CodeBlock from '@/app/common/CodeBlock.vue'
 import { fetchAllResources, getSome, getZoneDpServerAuthType } from '@/helpers'
 import Kuma from '@/services/kuma'
 import { getTableData } from '@/utils/tableDataUtils'
