+++ src/views/Onboarding/AddNewServicesCode.vue
@@ -83,8 +83,8 @@ import { KCard } from '@kong/kongponents'
 import Kuma from '@/services/kuma'
 import { PRODUCT_NAME } from '@/consts'
 import { kumaDpServerUrl } from '@/configUrl'
-import CodeBlock from '@/components/CodeBlock.vue'
-import LoadingBox from '@/components/LoadingBox.vue'
+import CodeBlock from '@/app/common/CodeBlock.vue'
+import LoadingBox from '@/app/common/LoadingBox.vue'
 import OnboardingNavigation from '@/views/Onboarding/components/OnboardingNavigation.vue'
 import OnboardingHeading from '@/views/Onboarding/components/OnboardingHeading.vue'
 import OnboardingPage from '@/views/Onboarding/components/OnboardingPage.vue'
