+++ src/views/Onboarding/DataplanesOverview.vue
@@ -59,7 +59,7 @@
 import { PRODUCT_NAME, OFFLINE } from '@/consts'
 import { getItemStatusFromInsight } from '@/dataplane'
 import Kuma from '@/services/kuma'
-import LoadingBox from '@/components/LoadingBox.vue'
+import LoadingBox from '@/app/common/LoadingBox.vue'
 import OnboardingNavigation from '@/views/Onboarding/components/OnboardingNavigation.vue'
 import OnboardingHeading from '@/views/Onboarding/components/OnboardingHeading.vue'
 import OnboardingPage from '@/views/Onboarding/components/OnboardingPage.vue'
