+++ src/views/Wizard/views/DataplaneUniversal.vue
@@ -404,7 +404,7 @@ import json2yaml from '@appscode/json2yaml'
 import Kuma from '@/services/kuma'
 import { kumaDpServerUrl } from '@/configUrl'
 import { kebabCase } from '@/helpers'
-import CodeBlock from '@/components/CodeBlock.vue'
+import CodeBlock from '@/app/common/CodeBlock.vue'
 import FormFragment from '@/views/Wizard/components/FormFragment.vue'
 import StepSkeleton from '@/views/Wizard/components/StepSkeleton.vue'
 import EnvironmentSwitcher from '@/views/Wizard/components/EnvironmentSwitcher.vue'
