+++ src/views/Wizard/views/Mesh.vue
@@ -591,7 +591,7 @@ import { mapGetters } from 'vuex'
 import Kuma from '@/services/kuma'
 import { rejectKeys } from '@/views/Wizard/helpers'
 import { kebabCase } from '@/helpers'
-import CodeBlock from '@/components/CodeBlock.vue'
+import CodeBlock from '@/app/common/CodeBlock.vue'
 import FormatForCLI from '@/mixins/FormatForCLI'
 import FormFragment from '@/views/Wizard/components/FormFragment.vue'
 import TabsWidget from '@/components/Utils/TabsWidget.vue'
